refactor: changes components naming to kebab case
